Move web state to separate folder
