+++ app-compose/src/main/kotlin/com/pitchedapps/frost/web/state/FrostWebState.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.pitchedapps.frost.web
+package com.pitchedapps.frost.web.state
 
 import mozilla.components.lib.state.State
