+++ app-compose/src/main/kotlin/com/pitchedapps/frost/web/state/FrostWebStore.kt
@@ -14,23 +14,31 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.pitchedapps.frost.web
+package com.pitchedapps.frost.web.state
 
-import com.google.common.flogger.FluentLogger
+import com.pitchedapps.frost.facebook.FB_URL_BASE
 import mozilla.components.lib.state.Middleware
-import mozilla.components.lib.state.MiddlewareContext
+import mozilla.components.lib.state.Store
 
-class FrostLoggerMiddleware(private val tag: String) : Middleware<FrostWebState, FrostAction> {
-  override fun invoke(
-    context: MiddlewareContext<FrostWebState, FrostAction>,
-    next: (FrostAction) -> Unit,
-    action: FrostAction
+/**
+ * See
+ * https://github.com/mozilla-mobile/firefox-android/blob/main/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/store/BrowserStore.kt
+ *
+ * For firefox example.
+ */
+class FrostWebStore(
+  tag: String,
+  initialState: FrostWebState = FrostWebState(),
+  middleware: List<Middleware<FrostWebState, FrostWebAction>> = emptyList(),
+) :
+  Store<FrostWebState, FrostWebAction>(
+    initialState,
+    FrostWebReducer::reduce,
+    middleware,
+    "FrostStore-$tag",
   ) {
-    logger.atInfo().log("FrostWebAction-%s: %s - %s", tag, action::class.simpleName, action)
-    next(action)
-  }
-
-  companion object {
-    private val logger = FluentLogger.forEnclosingClass()
+  init {
+    dispatch(InitAction)
+    dispatch(UserAction.LoadUrlAction(FB_URL_BASE))
   }
 }
